web_search/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
newspaper3k
beautifulsoup4
lxml
lxml_html_clean
stealth-requests
deep_translator
langdetect
htmldate